Move VerifiedOps to its own module.
