+++ core/src/mantle/traits/preverified_tx.rs
@@ -1,4 +1,4 @@
-use crate::mantle::{traits::mantle_tx::MantleTxWithProofs, transactions::tx::VerifiedOps};
+use crate::mantle::{traits::mantle_tx::MantleTxWithProofs, transactions::VerifiedOps};
 
 pub trait PreverifiedMantleTx: MantleTxWithProofs {
     /// Returns the cursor to the verified operations in this transaction.
